requirements.txt
openai
PyYaml
semgrep
networkx